requirements.txt
xrpl-py==4.3.1
base58==2.1.1
pynacl==1.6.1
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.5
nicegui==3.8.0